fix: resolve BOFUAnalyticsWrapper prop mismatch and checkout page prerender errors
Three BOFU SEO pages (best-dog-grooming-software, mobile-grooming-software,
moego-alternatives) passed invalid pageName/competitorName props to
BOFUAnalyticsWrapper instead of required sectionIds/sectionTitles. This
caused TypeScript errors and broke static prerendering.

Checkout pages (success, error, cancel) were use client pages using
useSearchParams which fail Next.js static prerendering. Split each into a
server component page.tsx (with force-dynamic) and a client component.

Also added @testing-library/dom to devDependencies (missing peer dep of
@testing-library/react v16).
