+++ package.json
@@ -33,6 +33,7 @@
   "devDependencies": {
     "@playwright/test": "^1.44.0",
     "@tailwindcss/postcss": "^4.2.2",
+    "@testing-library/dom": "^10.4.1",
     "@testing-library/jest-dom": "^6.9.1",
     "@testing-library/react": "^16.3.2",
     "@testing-library/user-event": "^14.6.1",
